+++ app/components/Routes/about/index.tsx
@@ -1,0 +1,16 @@
+import ReactDOMServer from "react-dom/server";
+import Header from "~/components/Layouts/Header";
+import Main from "~/components/Layouts/Main";
+import Content from "./parts/Content";
+import ContentMaterial from "../blogs/components/materials/Content";
+
+export default function About() {
+  return (
+    <>
+      <Header pageTitle="ABOUT" />
+      <Main>
+        <ContentMaterial content={ReactDOMServer.renderToString(Content())} />
+      </Main>
+    </>
+  );
+}
