+++ app/components/Routes/blogs/components/materials/Content.tsx
