+++ app/components/Routes/blogs/components/materials/Image.tsx
