+++ app/components/Routes/blogs/components/parts/Image/Figure.tsx
