+++ app/routes/blogs.$blogId.tsx
@@ -2,10 +2,8 @@ import type { LoaderArgs, V2_MetaFunction } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
 import { getBlogDetail, getMovieData } from "~/libs/micro-cms/client.server";
 import type { Blog, MovieData } from "~/libs/micro-cms/client.server";
-import Main from "~/components/Layouts/Main";
 import { domPurify } from "~/libs/sanitize/client.server";
-import Content from "./blogs/components/materials/Content";
-import BlogPageHeader from "./blogs/components/parts/Header";
+import BlogId from "~/components/Routes/blogs";
 
 type LoaderData = {
   blog: Blog;
@@ -63,31 +61,12 @@ export const meta: V2_MetaFunction<typeof loader> = ({ data }) => {
   ];
 };
 
-export default function BlogId() {
+export default function BlogDetail() {
   const { blog, movieDataList } = useLoaderData<LoaderData>();
   return (
-    <>
-      <BlogPageHeader
-        pageTitle={blog.title}
-        publishedAt={blog.publishedAt}
-      />
-      <Main>
-        <header className="title-slide">
-          <h3 className="font-bold text-center text-xs sm:text-sm mt-4 sm:mt-8 mb-8 sm:mb-16">
-            {blog.summery}
-          </h3>
-          <div
-            className={"h-48 sm:h-96 rounded-md"}
-            style={{
-              background: `linear-gradient(rgba(var(--color-overlay-rgb),0.5), rgba(var(--color-overlay-rgb),0.5)), url(${blog.eyecatch.url}) center / cover`,
-            }}
-          ></div>
-        </header>
-        <Content
-          content={blog.content}
-          movies={movieDataList}
-        />
-      </Main>
-    </>
+    <BlogId
+      blog={blog}
+      movieDataList={movieDataList}
+    />
   );
 }
